refactor: make clear separation of vanilly/soldering/vsss
