+++ Cargo.toml
@@ -17,9 +17,10 @@ description = "Garbled-circuit SNARK verifier using Groth16 (BN254), with reusab
 resolver = "2"
 
 [features]
-default = ["tracing_max_level", "tracing_scopes", "sp1-soldering", "test-utils"]
+default = ["tracing_max_level", "tracing_scopes", "vsss", "sp1-soldering", "test-utils"]
 sp1-soldering = ["sp1-prover", "sp1-core-machine", "sp1-core-executor", "sp1-primitives", "sp1-stark", "dirs", "bincode"]
 test-utils = []
+vsss = []
 
 tracing_max_level = ["tracing/max_level_trace", "tracing/release_max_level_info"]
 tracing_scopes = []
@@ -39,7 +40,6 @@ ark-std = "0.5.0"
 bitvec = "1.0.1"
 blake3 = { version = "1.6.1", features = ["traits-preview"] }
 circuit_component_macro = { path = "circuit_component_macro" }
-core_affinity = "0.8"
 crossbeam = "0.8"
 digest = "0.10"
 itertools = "0.14.0"
@@ -110,3 +110,7 @@ strip = "symbols"   # if you don't need debug symbols in release
 [[example]]
 name = "gsv_soldering"
 required-features = ["sp1-soldering"]
+
+[[example]]
+name = "gsv_vsss"
+required-features = ["vsss"]
