+++ src/gadgets/bn254/pairing.rs
@@ -181,156 +181,6 @@ fn g1_normalize_to_affine<C: CircuitContext>(circuit: &mut C, p: &G1Projective) 
     G1Projective { x, y, z }
 }
 
-/// Normalize a G2 projective point to affine (z = 1) in Montgomery domain.
-fn g2_normalize_to_affine<C: CircuitContext>(circuit: &mut C, q: &G2Projective) -> G2Projective {
-    // Convert projective (x, y, z) to affine (x/z^2, y/z^3, 1)
-    let inv_z = Fq2::inverse_montgomery(circuit, &q.z);
-    let inv_z2 = Fq2::square_montgomery(circuit, &inv_z);
-    let inv_z3 = Fq2::mul_montgomery(circuit, &inv_z2, &inv_z);
-
-    let x = Fq2::mul_montgomery(circuit, &q.x, &inv_z2);
-    let y = Fq2::mul_montgomery(circuit, &q.y, &inv_z3);
-    let one_c0 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ONE)).expect("const one");
-    let zero_c1 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ZERO)).expect("const zero");
-    let z = Fq2::from_components(one_c0, zero_c1);
-
-    G2Projective { x, y, z }
-}
-
-/// Apply Frobenius map to an affine G2 point: x' = (x^p) * c_x[i], y' = (y^p) * c_y[i], z = 1.
-/// Precondition: q.z encodes Montgomery ONE (affine form).
-#[allow(dead_code)]
-fn g2_frobenius_map_affine<C: CircuitContext>(
-    circuit: &mut C,
-    q: &G2Projective,
-    i: usize,
-) -> G2Projective {
-    // Reduce to cases used by BN254 Miller loop: i in {0,1,2}
-    let k = i % 3;
-
-    let one_c0 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ONE)).expect("const one");
-    let zero_c1 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ZERO)).expect("const zero");
-    let z_one = Fq2::from_components(one_c0, zero_c1);
-
-    match k {
-        0 => G2Projective {
-            x: q.x.clone(),
-            y: q.y.clone(),
-            z: z_one,
-        },
-        1 => {
-            let x_f1 = Fq2::frobenius_montgomery(circuit, &q.x, 1);
-            let y_f1 = Fq2::frobenius_montgomery(circuit, &q.y, 1);
-            let cx = ark_bn254::Config::TWIST_MUL_BY_Q_X;
-            let cy = ark_bn254::Config::TWIST_MUL_BY_Q_Y;
-            // Constants must be in Montgomery form for mul_by_constant_montgomery
-            let x_new = Fq2::mul_by_constant_montgomery(circuit, &x_f1, &Fq2::as_montgomery(cx));
-            let y_new = Fq2::mul_by_constant_montgomery(circuit, &y_f1, &Fq2::as_montgomery(cy));
-            G2Projective {
-                x: x_new,
-                y: y_new,
-                z: z_one,
-            }
-        }
-        _ => {
-            let x_f1 = Fq2::frobenius_montgomery(circuit, &q.x, 1);
-            let y_f1 = Fq2::frobenius_montgomery(circuit, &q.y, 1);
-            let cx = ark_bn254::Config::TWIST_MUL_BY_Q_X;
-            let cy = ark_bn254::Config::TWIST_MUL_BY_Q_Y;
-            let x1 = Fq2::mul_by_constant_montgomery(circuit, &x_f1, &Fq2::as_montgomery(cx));
-            let y1 = Fq2::mul_by_constant_montgomery(circuit, &y_f1, &Fq2::as_montgomery(cy));
-
-            let x_f2 = Fq2::frobenius_montgomery(circuit, &x1, 1);
-            let y_f2 = Fq2::frobenius_montgomery(circuit, &y1, 1);
-            let x2 = Fq2::mul_by_constant_montgomery(circuit, &x_f2, &Fq2::as_montgomery(cx));
-            let mut y2 = Fq2::mul_by_constant_montgomery(circuit, &y_f2, &Fq2::as_montgomery(cy));
-            y2 = Fq2::neg(circuit, y2);
-            G2Projective {
-                x: x2,
-                y: y2,
-                z: z_one,
-            }
-        }
-    }
-}
-
-/// Compute tangent line coefficients at R (variable-`Q` path) and advance R <- 2R.
-///
-/// Returns (R_next_affine, (c0, c1, c2)) where the line is represented as
-/// c0 * Y + c1 * X + c2 = 0 over Fq2, matching arkworks' BN254 prepared
-/// ordering (ell_0, ell_vw, ell_vv). Later evaluation multiplies c0 by P.y
-/// and c1 by P.x and uses a sparse Fq12 mul-by-034.
-#[allow(dead_code)]
-fn g2_line_coeffs_double<C: CircuitContext>(
-    circuit: &mut C,
-    r: &G2Projective,
-) -> (G2Projective, Fq6) {
-    // Normalize R to affine for simple slope/division formulas (a = 0 curve)
-    let r_aff = g2_normalize_to_affine(circuit, r);
-    let x = r_aff.x.clone();
-    let y = r_aff.y.clone();
-
-    // λ = (3*x^2) / (2*y)
-    let x2 = Fq2::square_montgomery(circuit, &x);
-    let three_x2 = Fq2::triple(circuit, &x2);
-    let two_y = Fq2::double(circuit, &y);
-    let inv_two_y = Fq2::inverse_montgomery(circuit, &two_y);
-    let lambda = Fq2::mul_montgomery(circuit, &three_x2, &inv_two_y);
-
-    // Line: Y - λ X - (y - λ x) = 0 ->
-    // c0 = 1, c1 = -λ, c2 = λ*x - y
-    let one_c0 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ONE)).expect("const one");
-    let zero_c1 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ZERO)).expect("const zero");
-    let c0 = Fq2::from_components(one_c0, zero_c1);
-    let c1 = Fq2::neg(circuit, lambda.clone());
-    let lambda_x = Fq2::mul_montgomery(circuit, &lambda, &x);
-    let c2 = Fq2::sub(circuit, &lambda_x, &y);
-
-    // Advance R <- 2R and normalize to affine for the caller
-    let r_next = crate::gadgets::bn254::g2::G2Projective::double_montgomery(circuit, r);
-    let r_next_aff = g2_normalize_to_affine(circuit, &r_next);
-
-    (r_next_aff, Fq6::from_components(c0, c1, c2))
-}
-
-/// Compute secant line coefficients through R and Q (variable-`Q` path) and advance R <- R + Q.
-///
-/// Returns (R_next_affine, (c0, c1, c2)) in the same representation order
-/// as `g2_line_coeffs_double`.
-#[allow(dead_code)]
-fn g2_line_coeffs_add<C: CircuitContext>(
-    circuit: &mut C,
-    r: &G2Projective,
-    q: &G2Projective,
-) -> (G2Projective, Fq6) {
-    // Normalize to affine to compute λ = (y_q - y_r) / (x_q - x_r)
-    let r_aff = g2_normalize_to_affine(circuit, r);
-    let q_aff = g2_normalize_to_affine(circuit, q);
-    let xr = r_aff.x.clone();
-    let yr = r_aff.y.clone();
-    let xq = q_aff.x.clone();
-    let yq = q_aff.y.clone();
-
-    let dy = Fq2::sub(circuit, &yq, &yr);
-    let dx = Fq2::sub(circuit, &xq, &xr);
-    let inv_dx = Fq2::inverse_montgomery(circuit, &dx);
-    let lambda = Fq2::mul_montgomery(circuit, &dy, &inv_dx);
-
-    // Line through R and Q: Y - λ X - (y_r - λ x_r) = 0
-    let one_c0 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ONE)).expect("const one");
-    let zero_c1 = Fq::new_constant(&Fq::as_montgomery(ark_bn254::Fq::ZERO)).expect("const zero");
-    let c0 = Fq2::from_components(one_c0, zero_c1);
-    let c1 = Fq2::neg(circuit, lambda.clone());
-    let lambda_xr = Fq2::mul_montgomery(circuit, &lambda, &xr);
-    let c2 = Fq2::sub(circuit, &lambda_xr, &yr);
-
-    // Advance R <- R + Q and normalize to affine for the caller
-    let r_next = crate::gadgets::bn254::g2::G2Projective::add_montgomery(circuit, r, q);
-    let r_next_aff = g2_normalize_to_affine(circuit, &r_next);
-
-    (r_next_aff, Fq6::from_components(c0, c1, c2))
-}
-
 impl FromWires for (G2Projective, Fq6) {
     fn from_wires(wires: &[crate::WireId]) -> Option<Self> {
         let (g2, fq6) = wires.split_at(G2Projective::ARITY);
@@ -1145,380 +995,6 @@ mod tests {
             });
 
         assert_eq!(result.output_value.value, expected_m);
-    }
-
-    #[test]
-    fn test_g2_normalize_to_affine_matches_host() {
-        // Generate random projective G2 point and expected affine (z=1) in Montgomery
-        let mut rng = ChaCha20Rng::seed_from_u64(0xA11FE2);
-        let r = ark_bn254::G2Projective::generator() * rnd_fr(&mut rng);
-        let r_aff = r.into_affine();
-        let expected = ark_bn254::G2Projective::new(r_aff.x, r_aff.y, ark_bn254::Fq2::ONE);
-        // Keep expected in canonical field representation for comparison with decoded canonical
-        let expected_std = expected;
-
-        // Circuit input containing the projective point (Montgomery)
-        struct InputG2norm {
-            p: ark_bn254::G2Projective,
-        }
-        struct WiresG2norm {
-            p: G2Projective,
-        }
-        impl CircuitInput for InputG2norm {
-            type WireRepr = WiresG2norm;
-            fn allocate(&self, mut issue: impl FnMut() -> WireId) -> Self::WireRepr {
-                WiresG2norm {
-                    p: G2Projective::new(&mut issue),
-                }
-            }
-            fn collect_wire_ids(repr: &Self::WireRepr) -> Vec<WireId> {
-                repr.p.to_wires_vec()
-            }
-        }
-        impl<M: CircuitMode<WireValue = bool>> EncodeInput<M> for InputG2norm {
-            fn encode(&self, repr: &Self::WireRepr, cache: &mut M) {
-                let p_m = G2Projective::as_montgomery(self.p);
-                let f = G2Projective::get_wire_bits_fn(&repr.p, &p_m).unwrap();
-                for &w in repr.p.to_wires_vec().iter() {
-                    if let Some(bit) = f(w) {
-                        cache.feed_wire(w, bit);
-                    }
-                }
-            }
-        }
-
-        // Output decoder for G2 projective wires -> ark type
-        struct OutG2norm {
-            val: ark_bn254::G2Projective,
-        }
-        impl CircuitOutput<ExecuteMode> for OutG2norm {
-            type WireRepr = G2Projective;
-            fn decode(wires: Self::WireRepr, cache: &mut ExecuteMode) -> Self {
-                // Decode Fq2 from wires helper
-                fn decode_fq2_from_wires(w: &Fq2, cache: &mut ExecuteMode) -> ark_bn254::Fq2 {
-                    // Read Montgomery limbs from wires, then convert back to standard form
-                    let c0_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(
-                        w.c0().0.clone(),
-                        cache,
-                    );
-                    let c1_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(
-                        w.c1().0.clone(),
-                        cache,
-                    );
-                    let c0_m = ark_bn254::Fq::from(c0_bi);
-                    let c1_m = ark_bn254::Fq::from(c1_bi);
-                    let fq2_m = ark_bn254::Fq2::new(c0_m, c1_m);
-                    Fq2::from_montgomery(fq2_m)
-                }
-                let x = decode_fq2_from_wires(&wires.x, cache);
-                let y = decode_fq2_from_wires(&wires.y, cache);
-                let z = decode_fq2_from_wires(&wires.z, cache);
-                OutG2norm {
-                    val: ark_bn254::G2Projective::new(x, y, z),
-                }
-            }
-        }
-
-        let input = InputG2norm { p: r };
-        let res =
-            CircuitBuilder::streaming_execute::<_, _, OutG2norm>(input, 20_000, |ctx, wires| {
-                g2_normalize_to_affine(ctx, &wires.p)
-            });
-
-        assert_eq!(res.output_value.val, expected_std);
-    }
-
-    // Local helper to decode Fq2 (Montgomery wires -> canonical field element)
-    fn decode_fq2_from_wires(w: &Fq2, cache: &mut ExecuteMode) -> ark_bn254::Fq2 {
-        let c0_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(w.0[0].0.clone(), cache);
-        let c1_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(w.0[1].0.clone(), cache);
-        let c0_m = ark_bn254::Fq::from(c0_bi);
-        let c1_m = ark_bn254::Fq::from(c1_bi);
-        let fq2_m = ark_bn254::Fq2::new(c0_m, c1_m);
-        Fq2::from_montgomery(fq2_m)
-    }
-
-    // Local helper to decode G2 projective wires into canonical host value
-    fn decode_g2proj_from_wires(
-        wires: &G2Projective,
-        cache: &mut ExecuteMode,
-    ) -> ark_bn254::G2Projective {
-        let x = decode_fq2_from_wires(&wires.x, cache);
-        let y = decode_fq2_from_wires(&wires.y, cache);
-        let z = decode_fq2_from_wires(&wires.z, cache);
-        ark_bn254::G2Projective::new(x, y, z)
-    }
-
-    #[test]
-    fn test_g2_line_coeffs_double_matches_host() {
-        let mut rng = ChaCha20Rng::seed_from_u64(0xD0u64);
-        let r = ark_bn254::G2Projective::generator() * rnd_fr(&mut rng);
-
-        // Host expected: coefficients from affine R and next = 2R (affine z=1)
-        let r_aff = r.into_affine();
-        let x = r_aff.x;
-        let y = r_aff.y;
-        let lambda = (x.square() + x.square() + x.square()) * (y.double()).inverse().unwrap();
-        let exp_c0 = ark_bn254::Fq2::ONE;
-        let exp_c1 = -lambda;
-        let exp_c2 = lambda * x - y;
-        let next_exp = (r.double()).into_affine();
-        let next_exp_proj =
-            ark_bn254::G2Projective::new(next_exp.x, next_exp.y, ark_bn254::Fq2::ONE);
-
-        // Execute gadget
-        struct Input {
-            r: ark_bn254::G2Projective,
-        }
-        struct Wires {
-            r: G2Projective,
-        }
-        impl CircuitInput for Input {
-            type WireRepr = Wires;
-            fn allocate(&self, mut issue: impl FnMut() -> WireId) -> Self::WireRepr {
-                Wires {
-                    r: G2Projective::new(&mut issue),
-                }
-            }
-            fn collect_wire_ids(repr: &Self::WireRepr) -> Vec<WireId> {
-                repr.r.to_wires_vec()
-            }
-        }
-        impl<M: CircuitMode<WireValue = bool>> EncodeInput<M> for Input {
-            fn encode(&self, repr: &Self::WireRepr, cache: &mut M) {
-                let r_m = G2Projective::as_montgomery(self.r);
-                let f = G2Projective::get_wire_bits_fn(&repr.r, &r_m).unwrap();
-                for &w in repr.r.to_wires_vec().iter() {
-                    if let Some(b) = f(w) {
-                        cache.feed_wire(w, b);
-                    }
-                }
-            }
-        }
-        struct Output {
-            next: ark_bn254::G2Projective,
-            coeffs: ark_bn254::Fq6,
-        }
-        impl CircuitOutput<ExecuteMode> for Output {
-            type WireRepr = (G2Projective, Fq6);
-            fn decode(wires: Self::WireRepr, cache: &mut ExecuteMode) -> Self {
-                let (next, c) = wires;
-                let c0 = decode_fq2_from_wires(&c.0[0], cache);
-                let c1 = decode_fq2_from_wires(&c.0[1], cache);
-                let c2 = decode_fq2_from_wires(&c.0[2], cache);
-                Self {
-                    next: decode_g2proj_from_wires(&next, cache),
-                    coeffs: ark_bn254::Fq6::new(c0, c1, c2),
-                }
-            }
-        }
-
-        let res = crate::circuit::CircuitBuilder::streaming_execute::<_, _, Output>(
-            Input { r },
-            50_000,
-            |ctx, w| {
-                let (next, c) = g2_line_coeffs_double(ctx, &w.r);
-                (next, c)
-            },
-        );
-
-        assert_eq!(res.output_value.next, next_exp_proj);
-        assert_eq!(res.output_value.coeffs.c0, exp_c0);
-        assert_eq!(res.output_value.coeffs.c1, exp_c1);
-        assert_eq!(res.output_value.coeffs.c2, exp_c2);
-    }
-
-    #[test]
-    fn test_g2_line_coeffs_add_matches_host() {
-        let mut rng = ChaCha20Rng::seed_from_u64(0xADDu64);
-        let r = ark_bn254::G2Projective::generator() * rnd_fr(&mut rng);
-        let q = ark_bn254::G2Projective::generator() * rnd_fr(&mut rng);
-
-        // Host expected from affine forms
-        let r_aff = r.into_affine();
-        let q_aff = q.into_affine();
-        // Avoid the rare vertical line; if so, tweak q by adding generator
-        let (x1, y1) = (r_aff.x, r_aff.y);
-        let (x2, y2) = (q_aff.x, q_aff.y);
-        let mut x2a = x2;
-        let mut y2a = y2;
-        if x2a == x1 {
-            // pathological, adjust deterministically
-            let g = ark_bn254::G2Affine::generator();
-            x2a = g.x;
-            y2a = g.y;
-        }
-        let lambda = (y2a - y1) * (x2a - x1).inverse().unwrap();
-        let exp_c0 = ark_bn254::Fq2::ONE;
-        let exp_c1 = -lambda;
-        let exp_c2 = lambda * x1 - y1;
-        let next_exp = (r + q).into_affine();
-        let next_exp_proj =
-            ark_bn254::G2Projective::new(next_exp.x, next_exp.y, ark_bn254::Fq2::ONE);
-
-        // Execute gadget
-        struct Input {
-            r: ark_bn254::G2Projective,
-            q: ark_bn254::G2Projective,
-        }
-        struct Wires {
-            r: G2Projective,
-            q: G2Projective,
-        }
-        impl CircuitInput for Input {
-            type WireRepr = Wires;
-            fn allocate(&self, mut issue: impl FnMut() -> WireId) -> Self::WireRepr {
-                Wires {
-                    r: G2Projective::new(&mut issue),
-                    q: G2Projective::new(issue),
-                }
-            }
-            fn collect_wire_ids(repr: &Self::WireRepr) -> Vec<WireId> {
-                let mut v = repr.r.to_wires_vec();
-                v.extend(repr.q.to_wires_vec());
-                v
-            }
-        }
-        impl<M: CircuitMode<WireValue = bool>> EncodeInput<M> for Input {
-            fn encode(&self, repr: &Self::WireRepr, cache: &mut M) {
-                let r_m = G2Projective::as_montgomery(self.r);
-                let f_r = G2Projective::get_wire_bits_fn(&repr.r, &r_m).unwrap();
-                for &w in repr.r.to_wires_vec().iter() {
-                    if let Some(b) = f_r(w) {
-                        cache.feed_wire(w, b);
-                    }
-                }
-                let q_m = G2Projective::as_montgomery(self.q);
-                let f_q = G2Projective::get_wire_bits_fn(&repr.q, &q_m).unwrap();
-                for &w in repr.q.to_wires_vec().iter() {
-                    if let Some(b) = f_q(w) {
-                        cache.feed_wire(w, b);
-                    }
-                }
-            }
-        }
-        struct Output {
-            next: ark_bn254::G2Projective,
-            coeffs: ark_bn254::Fq6,
-        }
-        impl CircuitOutput<ExecuteMode> for Output {
-            type WireRepr = (G2Projective, Fq6);
-            fn decode(wires: Self::WireRepr, cache: &mut ExecuteMode) -> Self {
-                let (next, c) = wires;
-                let c0 = decode_fq2_from_wires(&c.0[0], cache);
-                let c1 = decode_fq2_from_wires(&c.0[1], cache);
-                let c2 = decode_fq2_from_wires(&c.0[2], cache);
-                Self {
-                    next: decode_g2proj_from_wires(&next, cache),
-                    coeffs: ark_bn254::Fq6::new(c0, c1, c2),
-                }
-            }
-        }
-
-        let res = crate::circuit::CircuitBuilder::streaming_execute::<_, _, Output>(
-            Input { r, q },
-            80_000,
-            |ctx, w| {
-                let (next, c) = g2_line_coeffs_add(ctx, &w.r, &w.q);
-                (next, c)
-            },
-        );
-
-        assert_eq!(res.output_value.next, next_exp_proj);
-        assert_eq!(res.output_value.coeffs.c0, exp_c0);
-        assert_eq!(res.output_value.coeffs.c1, exp_c1);
-        assert_eq!(res.output_value.coeffs.c2, exp_c2);
-    }
-
-    #[test]
-    fn test_g2_frobenius_map_affine_matches_host() {
-        let mut rng = ChaCha20Rng::seed_from_u64(0xF10B1);
-        let r_aff = (ark_bn254::G2Projective::generator() * rnd_fr(&mut rng)).into_affine();
-        let r_aff_proj = ark_bn254::G2Projective::new(r_aff.x, r_aff.y, ark_bn254::Fq2::ONE);
-
-        // Build expected host result for i in {1,2}
-        for i in [1usize, 2usize] {
-            let x_p = r_aff.x.frobenius_map(1);
-            let y_p = r_aff.y.frobenius_map(1);
-            let cx = ark_bn254::Config::TWIST_MUL_BY_Q_X;
-            let cy = ark_bn254::Config::TWIST_MUL_BY_Q_Y;
-            let (exp_x, exp_y) = if i == 1 {
-                (x_p * cx, y_p * cy)
-            } else {
-                let x1 = x_p * cx;
-                let y1 = y_p * cy;
-                (x1.frobenius_map(1) * cx, -(y1.frobenius_map(1) * cy))
-            };
-
-            // Streaming execute: encode affine projective (z=1) and apply gadget
-            struct Input {
-                p: ark_bn254::G2Projective,
-            }
-            struct Wires {
-                p: G2Projective,
-            }
-            impl CircuitInput for Input {
-                type WireRepr = Wires;
-                fn allocate(&self, mut issue: impl FnMut() -> WireId) -> Self::WireRepr {
-                    Wires {
-                        p: G2Projective::new(&mut issue),
-                    }
-                }
-                fn collect_wire_ids(repr: &Self::WireRepr) -> Vec<WireId> {
-                    repr.p.to_wires_vec()
-                }
-            }
-            impl<M: CircuitMode<WireValue = bool>> EncodeInput<M> for Input {
-                fn encode(&self, repr: &Self::WireRepr, cache: &mut M) {
-                    let p_m = G2Projective::as_montgomery(self.p);
-                    let f = G2Projective::get_wire_bits_fn(&repr.p, &p_m).unwrap();
-                    for &w in repr.p.to_wires_vec().iter() {
-                        if let Some(bit) = f(w) {
-                            cache.feed_wire(w, bit);
-                        }
-                    }
-                }
-            }
-            struct OutXY {
-                x: ark_bn254::Fq2,
-                y: ark_bn254::Fq2,
-                z: ark_bn254::Fq2,
-            }
-            impl CircuitOutput<ExecuteMode> for OutXY {
-                type WireRepr = G2Projective;
-                fn decode(wires: Self::WireRepr, cache: &mut ExecuteMode) -> Self {
-                    fn decode_fq2_from_wires(w: &Fq2, cache: &mut ExecuteMode) -> ark_bn254::Fq2 {
-                        let c0_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(
-                            w.c0().0.clone(),
-                            cache,
-                        );
-                        let c1_bi = <BigUintOutput as CircuitOutput<ExecuteMode>>::decode(
-                            w.c1().0.clone(),
-                            cache,
-                        );
-                        Fq2::from_montgomery(ark_bn254::Fq2::new(
-                            ark_bn254::Fq::from(c0_bi),
-                            ark_bn254::Fq::from(c1_bi),
-                        ))
-                    }
-                    let x = decode_fq2_from_wires(&wires.x, cache);
-                    let y = decode_fq2_from_wires(&wires.y, cache);
-                    let z = decode_fq2_from_wires(&wires.z, cache);
-                    OutXY { x, y, z }
-                }
-            }
-
-            let input = Input { p: r_aff_proj };
-            let result =
-                CircuitBuilder::streaming_execute::<_, _, OutXY>(input, 20_000, |ctx, wires| {
-                    g2_frobenius_map_affine(ctx, &wires.p, i)
-                });
-
-            let out = result.output_value;
-            assert_eq!(out.x, exp_x);
-            assert_eq!(out.y, exp_y);
-            assert_eq!(out.z, ark_bn254::Fq2::ONE);
-        }
     }
 
     #[test]
